+++ packages/google_atlas/test/widget_tests/google_atlas_widget_test.dart
@@ -684,6 +684,7 @@ class GoogleAtlasProviderSample extends Provider {
   @override
   Widget build({
     CameraPosition initialCameraPosition,
+    CameraPosition currentCameraPosition,
     Set<Marker> markers,
     Set<Circle> circles,
     Set<Polygon> polygons,