+++ packages/atlas/CHANGELOG.md
@@ -1,5 +1,11 @@
 # atlas
 
+## v0.8.0 (2020-7-9)
+
+- `currentCameraPosition` added to Atlas Map
+
+# atlas
+
 ## v0.7.0 (2020-5-12)
 
 - Added `changeUserLocationIcon` function